# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Endpoints for authentication, login and register\n\n\u003cdetails open\u003e\n\u003csummary\u003e\u003ca href=\"javascript:void(0)\" style=\"cursor: pointer !important;\"\u003e📋 Message Codes\u003c/a\u003e\u003c/summary\u003e\n\n| Code   | Description                   |\n| ------ | ------------------------------|\n| 1000   | successful                    |\n| 1001   | unsuccessful                  |\n| 1002   | found                         |\n| 1003   | not found                     |\n| 4000   | internal error                |\n| 4001   | invalid request body          |\n| 4002   | invalid authentication token  |\n| 4003   | missing authentication token  |\n| 4004   | invalid signature             |\n| 4005   | missing signature             |\n| 4006   | invalid appid                 |\n| 4007   | missing appid                 |\n| 4008   | invalid query parameters      |\n\n\u003c/details\u003e",
        "title": "LBE API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/pin": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the burn PIN for a given email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user burn PIN",
                "parameters": [
                    {
                        "description": "Email + new burn PIN",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateBurnPin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update successful",
                        "schema": {
                            "$ref": "#/definitions/responses.UpdateUserSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/update/{external_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates a user's profile fields (non‐zero values in the JSON body).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user external ID",
                        "name": "external_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Profile fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update successful",
                        "schema": {
                            "$ref": "#/definitions/responses.UpdateUserSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "existing user not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{external_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the profile (including phone numbers) for a given user by external_id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user external ID",
                        "name": "external_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user found",
                        "schema": {
                            "$ref": "#/definitions/responses.GetUserSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing external_id path parameter",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "existing user not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "requests.PhoneNumber": {
            "type": "object",
            "properties": {
                "phone_number": {
                    "description": "Phone number (digits only)",
                    "type": "string",
                    "example": "1234123123"
                },
                "phone_type": {
                    "description": "Type of phone number: mobile, office, home, fax, other",
                    "type": "string",
                    "example": "home"
                },
                "preference_flags": {
                    "description": "Flags like \"primary\"",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"primary\"]"
                    ]
                }
            }
        },
        "requests.Referral": {
            "type": "object",
            "properties": {
                "referral_code": {
                    "description": "Code of the referring customer",
                    "type": "string",
                    "example": "JOHN-70A756"
                }
            }
        },
...
        "requests.UpdateBurnPin": {
            "type": "object",
            "required": [
                "burn_pin",
                "email"
            ],
            "properties": {
                "burn_pin": {
                    "description": "BurnPin is the new numeric PIN to set.",
                    "type": "integer",
                    "example": 4321
                },
                "email": {
                    "description": "Email of the user whose burn PIN is being updated.",
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "requests.UserRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address line 1",
                    "type": "string",
                    "example": "7 Tremont Street"
                },
                "address2": {
                    "description": "Address line 2",
                    "type": "string",
                    "example": "8 Tremont Street"
                },
                "city": {
                    "description": "City of residence",
                    "type": "string",
                    "example": "Boston"
                },
                "country": {
                    "description": "3-letter ISO-3166-1 country code",
                    "type": "string",
                    "example": "USA"
                },
                "dob": {
                    "description": "Date of birth, YYYY-MM-DD",
                    "type": "string",
                    "example": "1980-01-01"
                },
                "email": {
                    "description": "Customer’s email (encrypted at rest). Required if ExternalID is not specified.",
                    "type": "string",
                    "example": "john.smith@test.com"
                },
                "external_id": {
                    "description": "Identifier for customer in external system.\nRequired if Email is not specified.",
                    "type": "string",
                    "example": "1284111"
                },
                "external_id_type": {
                    "description": "Type associated with external identifier.",
                    "type": "string",
                    "example": "facebook"
                },
                "first_name": {
                    "description": "First name",
                    "type": "string",
                    "example": "John"
                },
                "gender": {
                    "description": "Gender of customer: m or f",
                    "type": "string",
                    "example": "m"
                },
                "ip": {
                    "description": "Customer’s IP at registration.",
                    "type": "string",
                    "example": "203.0.113.42"
                },
                "last_name": {
                    "description": "Last name (surname)",
                    "type": "string",
                    "example": "Smith"
                },
                "locale": {
                    "description": "Preferred locale, e.g. en-US",
                    "type": "string",
                    "example": "en-US"
                },
                "opted_in": {
                    "description": "Indicates opt-in to loyalty program (defaults to true if not set).",
                    "type": "boolean",
                    "example": true
                },
                "phone_numbers": {
                    "description": "Phone numbers associated with customer",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.PhoneNumber"
                    }
                },
                "referral": {
                    "description": "Referral info when signing up with a referrer code",
                    "allOf": [
                        {
                            "$ref": "#/definitions/requests.Referral"
                        }
                    ]
                },
                "state": {
                    "description": "State/province/region",
                    "type": "string",
                    "example": "MA"
                },
                "zip": {
                    "description": "Zip or postal code",
                    "type": "string",
                    "example": "02021"
                }
            }
        },
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is your internal API status code, e.g. 1002",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "description": "Message is a human‑readable description, e.g. \"invalid json request body\"",
                    "type": "string"
                }
            }
        },
        "responses.GetUserSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1002
                },
                "data": {
                    "$ref": "#/definitions/responses.User"
                },
                "message": {
                    "type": "string",
                    "example": "user found"
                }
            }
        },
...
        "responses.Identifier": {
            "type": "object",
            "properties": {
                "external_id": {
                    "type": "string"
                },
                "external_id_type": {
                    "type": "string"
                }
            }
        },
...
        "responses.PhoneNumber": {
            "type": "object",
            "properties": {
                "phone_number": {
                    "type": "string"
                },
                "phone_type": {
                    "type": "string"
                },
                "preference_flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "verified_ownership": {
                    "type": "boolean"
                }
            }
        },
...
        "responses.UpdateUserSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1000
                },
                "data": {
                    "$ref": "#/definitions/responses.User"
                },
                "message": {
                    "type": "string",
                    "example": "update successful"
                }
            }
        },
        "responses.User": {
            "type": "object",
            "properties": {
                "account_status": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "address2": {
                    "type": "string"
                },
                "auth_token": {
                    "type": "string"
                },
                "available_points": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "description": "format: \"2006-01-02 15:04:05\"",
                    "type": "string"
                },
                "dob": {
                    "description": "format: \"2006-01-02\"",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identifiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.Identifier"
                    }
                },
                "last_name": {
                    "type": "string"
                },
                "opted_in": {
                    "type": "boolean"
                },
                "phone_numbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.PhoneNumber"
                    }
                },
                "proxy_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "referrer_code": {
                    "type": "string"
                },
                "registered_at": {
                    "description": "same format as CreatedAt",
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "suspended": {
                    "type": "boolean"
                },
                "tier": {
                    "type": "string"
                },
                "updated_at": {
                    "description": "same format as CreatedAt",
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer \u003cyour-jwt\u003e\" to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}